+++ app/src/components/items/overview/metadata/overview.tsx
@@ -12,40 +12,7 @@ import {
 } from "@nypl/design-system-react-components";
 import { Fragment } from "react";
 import parse from "html-react-parser";
-
-const metadataFieldToDisplay = {
-  title: "Title",
-  collection: "Collection",
-  names: "Names",
-  origin: "Date / Origin",
-  tableOfContents: "Table Of Contents",
-  locations: "Library Location",
-  subjects: "Subjects",
-  genres: "Genres",
-  notes: "Notes",
-  physicalDescription: "Physical Description",
-  abstract: "Abstract",
-  languages: "Languages",
-  link: "Link",
-  identifiers: "Identifiers",
-  access: "Access",
-  rights: "Rights",
-  typeOfResource: "Type Of Resource",
-  dateCreated: "Date Created",
-  dateIssued: "Date Issued",
-};
-
-// Helper function to parse anchor tag from HTML string
-// annoyingly redundant
-const parseHtmlString = (html) => {
-  const parser = new DOMParser();
-  const doc = parser.parseFromString(html, "text/html");
-  const anchor = doc.querySelector("a");
-  return {
-    href: anchor?.getAttribute("href"),
-    text: anchor?.textContent,
-  };
-};
+import { metadataFieldToDisplay } from "@/src/utils/metadata/filterRenderableMetadata";
 
 // Designs use DS Link component but I wonder if we can get away with not using them for the
 // Items page since the manifests are returning the links
@@ -53,8 +20,6 @@ const StructuredCollectionsList = (rawCollections) => {
   return (
     <>
       {rawCollections.map((htmlString, index) => {
-        console.log("htmlString is: ", htmlString);
-        const { href, text } = parseHtmlString(htmlString);
         const key = `metadata-collection-${index}`;
         return (
           <span key={key} style={{ paddingLeft: `${(index - 1) * 1.5}rem` }}>
@@ -62,10 +27,6 @@ const StructuredCollectionsList = (rawCollections) => {
               <Icon name="navigationSubdirectoryArrowRight" size="large" />
             )}
             {parse(htmlString)}
-            {/* TODO: keeping this in here if we need to use the DS Link component */}
-            {/* <Link href={href}>
-              {text}
-            </Link> */}
             <br></br>
           </span>
         );
@@ -75,41 +36,36 @@ const StructuredCollectionsList = (rawCollections) => {
 };
 
 const MetadataOverview = ({ metadata }) => {
+  const typedMetadata = metadata as Record<string, string>;
+
   return (
     <>
       <Box>
-        {/* <Flex marginTop="m" marginBottom="m" flexDir="column"> */}
         <Heading size="heading6" marginBottom="xs">
           Item data
         </Heading>
-
-        {Object.keys(metadata)?.map((field, index) => {
-          const value = metadata[field];
-          if (value !== "" && metadataFieldToDisplay[field] !== "") {
-            if (field === "collection") {
-              const collections = value.split("<br>");
-              console.log("collections are: ", collections);
-              return (
-                <Fragment key="metadata-collection">
-                  <Text size="overline1" marginBottom="xs">
-                    {metadataFieldToDisplay[field]}
-                  </Text>
-                  <Text marginBottom="m">
-                    {StructuredCollectionsList(collections)}
-                  </Text>
-                </Fragment>
-              );
-            } else {
-              return (
-                <Fragment key={`"metadata-${field}`}>
-                  <Text size="overline1" marginBottom="xs">
-                    {metadataFieldToDisplay[field]}
-                  </Text>
-                  <Text marginBottom="m">{parse(value)}</Text>
-                </Fragment>
-              );
-            }
+        {Object.entries(typedMetadata).map(([field, value]) => {
+          if (field === "collection") {
+            const collections = value.split("<br>");
+            return (
+              <Fragment key="metadata-collection">
+                <Text size="overline1" marginBottom="xs">
+                  {metadataFieldToDisplay[field]}
+                </Text>
+                <Text marginBottom="m">
+                  {StructuredCollectionsList(collections)}
+                </Text>
+              </Fragment>
+            );
           }
+          return (
+            <Fragment key={`metadata-${field}`}>
+              <Text size="overline1" marginBottom="xs">
+                {metadataFieldToDisplay[field]}
+              </Text>
+              <Text marginBottom="m">{parse(value)}</Text>
+            </Fragment>
+          );
         })}
       </Box>
     </>