+++ app/src/models/item.ts
@@ -1,4 +1,19 @@
 import { Maniiifest } from "maniiifest";
+import {
+  convertManiiifestMetadata,
+  parseManifestMetadata,
+} from "../utils/metadata/parseManifestMetadata";
+import { normalizeItemMetadataFromManifest } from "../utils/metadata/normalizeItemMetdata";
+import { NormalizedItemMetadata } from "../types/NormalizedItemMetadata";
+import { getRenderableMetadata } from "../utils/metadata/filterRenderableMetadata";
+import {
+  extractFirstHrefFromHTML,
+  extractAllAnchorsFromHTML,
+} from "../utils/metadata/extractAnchorHrefs";
+import {
+  generateCitations,
+  CitationOutput,
+} from "../utils/metadata/generateCitations";
 
 // https://github.com/jptmoore/maniiifest
 // other resources:
@@ -8,92 +23,70 @@ import { Maniiifest } from "maniiifest";
 // TO DO: add isCartographic for map stuff
 export class ItemModel {
   uuid: string;
-  mods: any;
-  capture: any;
-  typeOfResource: string;
-  title: string;
-  isSingleCapture: boolean;
-  imageIDs?: string[];
-  href: string;
-  contentType: string;
-  hasItems?: boolean;
-  manifestURL: string;
   link: string;
+  manifestURL: string;
+  hasItems: boolean;
+  imageIDs: string[] | null;
+  metadata: NormalizedItemMetadata;
+  renderableMetadata: Record<string, string>;
+  title: string;
+  typeOfResource: string;
   isRestricted: boolean;
-  isImage: boolean;
-  archivesLink?: string | null | undefined;
-  catalogLink?: string | null | undefined;
   divisionLink: string;
-  metadata?: {
-    title: string;
-    names?: string;
-    collection?: string;
-    origin: string;
-    dateCreated?: string;
-    dateIssued: string;
-    tableOfContents?: string;
-    locations: string;
-    subjects?: string;
-    genres: string;
-    notes?: string;
-    physicalDescription?: string;
-    abstract?: string;
-    languages: string;
-    link?: string;
-    identifiers: string;
-    access: string;
-    rights: string;
-    typeOfResource: string;
-  };
+  contentType: string;
+  isImage: boolean;
+  archivesLink: string | null;
+  catalogLink: string | null;
+  citationData: CitationOutput;
+  breadcrumbData: any;
 
   constructor(uuid: string, manifest: any) {
-    // get custom label from manifest
     const parser = new Maniiifest(manifest);
-    this.hasItems = manifest.items.length > 0 ? true : false;
-    const canvases = Array.from(parser.iterateManifestCanvas());
-    const canvasAnnotations = Array.from(
-      parser.iterateManifestCanvasAnnotation()
-    );
-    console.log("canvasAnnotations is: ", canvasAnnotations);
-
-    // example canvas.id is: "https://iiif.nypl.org/iiif/3/TH-38454/full/!700,700/0/default.jpg"
-    this.imageIDs = this.hasItems
-      ? canvases.map((canvas) => {
-          return canvas.id.split("/")[5];
-        })
-      : [];
-
-    // const label = parser?.getManifestLabelByLanguage("en");
-    const metadata = Array.from(parser.iterateManifestMetadata());
-    const manifestMetadataHash = {};
-
-    if (metadata) {
-      for (const field of metadata) {
-        const label = field.label["en"][0];
-        const value = field.value["en"];
-        manifestMetadataHash[label] = value;
-      }
-    }
-
+    // Non-Manifest/Metadata related fields
     this.uuid = uuid;
-    this.link =
-      process.env.APP_ENV === "development" || process.env.APP_ENV === "qa"
-        ? `https://qa-digitalcollections.nypl.org/items/${this.uuid}`
-        : `https://digitalcollections.nypl.org/items/${this.uuid}`;
-    this.typeOfResource = manifestMetadataHash["Resource Type"]
-      ? manifestMetadataHash["Resource Type"].toString()
+    this.manifestURL = `${process.env.COLLECTIONS_API_URL}/manifests/${uuid}`;
+
+    // Manifest related fields
+    this.hasItems = manifest.items.length > 0;
+    const canvases = Array.from(parser.iterateManifestCanvas());
+
+    // Metadata assignment
+    const manifestMetadataArray = Array.from(parser.iterateManifestMetadata());
+    // convert Maniifest MetadataT to custom type
+    const convertedManifestMetadata = convertManiiifestMetadata(
+      manifestMetadataArray
+    );
+    const rawManifestMetadata = parseManifestMetadata(
+      convertedManifestMetadata
+    );
+    const normalizedManifestMetadata =
+      normalizeItemMetadataFromManifest(rawManifestMetadata);
+
+    this.metadata = normalizedManifestMetadata;
+    this.renderableMetadata = getRenderableMetadata(this.metadata);
+
+    // Manifest Metadata related fields
+    this.title = rawManifestMetadata["Title"]
+      ? rawManifestMetadata["Title"].toString()
       : "";
-    this.title = manifestMetadataHash["Title"]
-      ? manifestMetadataHash["Title"].toString()
+
+    this.typeOfResource = rawManifestMetadata["Resource Type"]
+      ? rawManifestMetadata["Resource Type"].toString()
       : "";
-    this.isRestricted = manifestMetadataHash["Is Restricted"]
-      ? manifestMetadataHash["Is Restricted"].toString().toLowerCase() ===
-        "true"
+
+    this.isRestricted = rawManifestMetadata["Is Restricted"]
+      ? rawManifestMetadata["Is Restricted"].toString().toLowerCase() === "true"
       : true;
 
+    //this will break in Prod if we don't deploy API first bc the name of the field is "Library Location"
+    this.divisionLink =
+      this.isRestricted && rawManifestMetadata["Division"]
+        ? rawManifestMetadata["Division"].toString()
+        : rawManifestMetadata["Library Locations"][0] || "";
+
     // for viewer configs and order print button
-    this.contentType = manifestMetadataHash["Content Type"]
-      ? manifestMetadataHash["Content Type"][0].toString()
+    this.contentType = rawManifestMetadata["Content Type"]
+      ? rawManifestMetadata["Content Type"][0].toString()
       : "";
 
     // only used for order print button
@@ -103,80 +96,65 @@ export class ItemModel {
         this.contentType === "single image" ||
         this.contentType === "multiple images");
 
-    this.divisionLink =
-      this.isRestricted && manifestMetadataHash["Division"]
-        ? manifestMetadataHash["Division"].toString()
-        : manifestMetadataHash["Library Locations"][0];
+    // get list of imageIDs for order print button
+    // example canvas.id is: "https://iiif.nypl.org/iiif/3/TH-38454/full/!700,700/0/default.jpg"
+    this.imageIDs =
+      this.hasItems && this.isImage
+        ? canvases.map((canvas) => {
+            return canvas.id.split("/")[5];
+          })
+        : null;
 
-    this.manifestURL = `${process.env.COLLECTIONS_API_URL}/manifests/${uuid}`;
+    // Special NYPL Identifiers for external links
+    const identifiers = rawManifestMetadata["Identifiers"];
 
-    this.metadata = {
-      title: manifestMetadataHash["Title"]
-        ? manifestMetadataHash["Title"].toString()
-        : "",
-      collection: manifestMetadataHash["Collection"]
-        ? manifestMetadataHash["Collection"].join("<br>") //TODO: add arrows
-        : "",
-      names: manifestMetadataHash["Names"]
-        ? manifestMetadataHash["Names"].join("<br>")
-        : "",
-      origin: manifestMetadataHash["Dates / origin"]
-        ? manifestMetadataHash["Dates / origin"].join("<br>")
-        : "",
-      dateIssued: manifestMetadataHash["Date Issued"]
-        ? manifestMetadataHash["Date Issued"].toString()
-        : "",
-      tableOfContents: manifestMetadataHash["Table of Contents"]
-        ? manifestMetadataHash["Table of Contents"].toString()
-        : "",
-      locations: manifestMetadataHash["Library Locations"]
-        ? manifestMetadataHash["Library Locations"].join("<br>")
-        : "",
-      subjects: manifestMetadataHash["Subjects"]
-        ? manifestMetadataHash["Subjects"].join("<br>")
-        : "",
-      genres: manifestMetadataHash["Genres"]
-        ? manifestMetadataHash["Genres"].join("<br>")
-        : "",
-      notes: manifestMetadataHash["Notes"]
-        ? manifestMetadataHash["Notes"].join("")
-        : "",
-      physicalDescription: manifestMetadataHash["Physical Description"]
-        ? manifestMetadataHash["Physical Description"].join("<br>")
-        : "",
-      typeOfResource: manifestMetadataHash["Resource Type"]
-        ? manifestMetadataHash["Resource Type"].join("<br>")
-        : "",
-      abstract: manifestMetadataHash["Abstract"]
-        ? manifestMetadataHash["Abstract"].toString()
-        : "",
-      languages: manifestMetadataHash["Languages"]
-        ? manifestMetadataHash["Lanugages"].toString()
-        : "",
-      link: manifestMetadataHash["Link"]
-        ? manifestMetadataHash["Link"].toString()
-        : "",
-      identifiers: manifestMetadataHash["Identifiers"]
-        ? manifestMetadataHash["Identifiers"].join("<br>")
-        : "",
-      access: manifestMetadataHash["Access"]
-        ? manifestMetadataHash["Access"].toString()
-        : "",
-      rights: manifestMetadataHash["Rights"]
-        ? manifestMetadataHash["Rights"].toString()
-        : "",
-    };
+    const archivesLink = identifiers.find((identifier) =>
+      identifier.includes("Archives ID")
+    );
 
-    const identifiers = manifestMetadataHash["Identifiers"];
     const catalogLink = identifiers.find((identifier) =>
       identifier.includes("NYPL Catalog ID (bnumber)")
     );
 
-    this.catalogLink = catalogLink ? catalogLink : "";
-    const archivesLink = identifiers.find((identifier) => {
-      identifiers.includes("Archives ID");
+    this.archivesLink = archivesLink
+      ? extractFirstHrefFromHTML(archivesLink)
+      : null;
+
+    this.catalogLink = catalogLink
+      ? extractFirstHrefFromHTML(catalogLink)
+      : null;
+
+    // Citation Data
+    this.citationData = generateCitations({
+      title: this.title,
+      link: this.link,
+      location:
+        extractAllAnchorsFromHTML(
+          this.metadata?.locations?.split("<br>")[0] ?? ""
+        )[0].text ?? "",
+      resource:
+        extractAllAnchorsFromHTML(
+          this.metadata?.typeOfResource?.split("<br>")[0] ?? ""
+        )[0].text ?? "",
+      origin: this.metadata?.origin,
+      dateIssued: this.metadata?.dateIssued,
     });
 
-    this.archivesLink = archivesLink ? archivesLink : "";
+    // Breadcrumb Data
+    const divisionLinkObj = extractAllAnchorsFromHTML(
+      this.metadata?.locations?.split("<br>")[0] ?? ""
+    )[0];
+    divisionLinkObj["path"] = new URL(divisionLinkObj.href).pathname;
+
+    // note: this points to the top level collection, not the immediate parent collection or subcollection
+    const collectionLinkObj = extractAllAnchorsFromHTML(
+      this.metadata?.collection?.split("<br>")[0] ?? ""
+    )[0];
+    collectionLinkObj["path"] = new URL(collectionLinkObj.href).pathname;
+
+    this.breadcrumbData = {
+      division: divisionLinkObj,
+      collection: collectionLinkObj,
+    };
   }
 }